+++ tests/test_04_dxf_high_level_structs/test_406_blocks_section.py
@@ -6,6 +6,7 @@ import ezdxf
 from ezdxf.tools.test import load_entities
 from ezdxf.sections.blocks import BlocksSection
 from ezdxf.lldxf.tagwriter import TagCollector
+from ezdxf.entities import factory
 
 
 @pytest.fixture
@@ -15,11 +16,21 @@ def dxf12():
 
 @pytest.fixture
 def blocks(dxf12):
-    return BlocksSection(dxf12, list(load_entities(TESTBLOCKS, 'BLOCKS', dxf12)))
+    return BlocksSection(dxf12,
+                         list(load_entities(TESTBLOCKS, 'BLOCKS', dxf12)))
+
+
+@pytest.fixture
+def doc():
+    doc = ezdxf.new()
+    doc.blocks.new('_ARCHTICK')
+    doc.blocks.new('_OPEN30')
+    return doc
 
 
 def test_empty_section(dxf12):
-    blocks = BlocksSection(dxf12, list(load_entities(EMPTYSEC, 'BLOCKS', dxf12)))
+    blocks = BlocksSection(dxf12,
+                           list(load_entities(EMPTYSEC, 'BLOCKS', dxf12)))
     # the NES creates automatically *Model_Space and *Paper_Space blocks
     assert '*Model_Space' in blocks
     assert '*Paper_Space' in blocks
@@ -33,7 +44,8 @@ def test_empty_section(dxf12):
     # tag[3] is a arbitrary handle
     assert collector.tags[3][0] == 5
     assert collector.tags[4] == (8, '0')  # default layer '0'
-    assert collector.tags[5] == (2, '$Model_Space')  # export modelspace with leading '$' for R12
+    assert collector.tags[5] == (
+        2, '$Model_Space')  # export modelspace with leading '$' for R12
     assert collector.tags[-1] == (0, 'ENDSEC')
 
 
@@ -120,24 +132,50 @@ def test_safe_delete_block(blocks, dxf12):
         blocks.delete_block('_ArchTick', safe=True)
 
 
-def test_delete_all_blocks(blocks):
-    blocks.delete_all_blocks(safe=False)
-    blocks = list(blocks)
-    # assure not deleting layout blocks
-    assert len(blocks) == 2
-    block_names = [block.name for block in blocks]
-    block_names.sort()
-    assert ['*Model_Space', '*Paper_Space'] == block_names
+def test_do_not_delete_layouts_and_special_arrow_blocks(doc):
+    doc.blocks.delete_all_blocks()
+    assert len(doc.blocks) == 4
+    block_names = set(block.name for block in doc.blocks)
+    assert block_names == {'*Model_Space', '*Paper_Space', '_ARCHTICK',
+                           '_OPEN30'}
 
 
-def test_safe_delete_all_blocks(blocks):
-    blocks.delete_all_blocks(safe=True)
-    blocks = list(blocks)
-    # assure not deleting layout blocks
-    assert len(blocks) == 4
-    block_names = [block.name for block in blocks]
-    block_names.sort()
-    assert ['*Model_Space', '*Paper_Space', '_ARCHTICK', '_OPEN30'] == block_names
+def test_do_not_purge_layout_and_special_arrow_blocks(doc):
+    doc.blocks.purge()
+    assert len(doc.blocks) == 4
+    block_names = set(block.name for block in doc.blocks)
+    assert block_names == {'*Model_Space', '*Paper_Space', '_ARCHTICK',
+                           '_OPEN30'}
+
+
+@pytest.mark.parametrize('name', ['*D1', '*T1', '*U1', '*X1', '*E1', '*A1'])
+def test_do_purge_unused_special_blocks(name, doc):
+    doc.blocks.new(name)
+    doc.blocks.purge()
+    assert name not in doc.blocks
+
+
+@pytest.mark.parametrize('name', ['*U1', '*X1', '*E1', '*A1'])
+def test_do_not_purge_referenced_special_blocks(name, doc):
+    doc.blocks.new(name)
+    msp = doc.modelspace()
+    msp.add_blockref(name, (0, 0))
+    doc.blocks.purge()
+    assert name in doc.blocks
+
+
+def test_do_not_purge_used_dimension(doc):
+    # DIMENSION (and ACAD_TABLE) block are not referenced by explicit INSERT
+    # entity.
+    dimension = doc.dxffactory.create_db_entry(
+        'DIMENSION',
+        dxfattribs={'geometry': '*D01'},
+    )
+    msp = doc.modelspace()
+    msp.add_entity(dimension)
+    doc.blocks.new('*D01')
+    doc.blocks.purge()
+    assert '*D01' in doc.blocks
 
 
 def test_rename_block(blocks):